Add not_found label to portfolio tag taxonomy.
This label, added in WP 3.6, is the text displayed via clicking 'Choose from the most used tags' in the taxonomy meta box when no terms are available. This string isn't used on hierarchical taxonomies, but I've included it anyway for parity between the taxonomies label arrays, the same as parent_item is included for both taxonomies.

I've also rearranged the label arrays into a slightly more logical order:

 * Names
 * Strings with singular references (uppercase)
 * Strings with plural references (uppercase)
 * Strings with plural references (lowercase)